+++ projects/atft/package.json
@@ -1,6 +1,6 @@
 {
   "name": "atft",
-  "version": "1.3.4",
+  "version": "1.4.0",
   "description": "Ready-to-use Angular components for Three.js, see https://makimenko.github.io/angular-template-for-threejs/.",
   "keywords": [
     "threejs",
