+++ src/app/app.component.html
@@ -1,1 +1,49 @@
-<app-editor-layout></app-editor-layout>
+<atft-orbit-controls [rotateSpeed]=1 [zoomSpeed]=1.2>
+    <atft-renderer-canvas>
+
+      <atft-perspective-camera [fov]=60 [near]=1 [far]=1100 [positionX]=cameraPositionX positionY=50 positionZ=50
+                               atft-raycaster-camera atft-raycaster-enable>
+      </atft-perspective-camera>
+
+    <atft-scene>
+      <atft-axes-helper size=200></atft-axes-helper>
+      <atft-grid-helper size=100 divisions=10></atft-grid-helper>
+
+      <atft-point-light color="0xffffff" intensity="0.9" distance="1000" translateX=50 translateY=50 translateZ=50></atft-point-light>
+      <atft-point-light color="0xffffff" intensity="0.1" distance="1000" translateX=-50 translateY=50 translateZ=20></atft-point-light>
+
+      <atft-obj-loader
+        model="assets/model/smiley/smiley.obj"
+        material="assets/model/smiley/smiley.mtl"
+        texturePath="assets/model/smiley/"
+        rotateX={{-90|deg2rad}}
+        translateX=-20
+        translateY=3>
+      </atft-obj-loader>
+
+      <div *ngFor="let z of [-10, -5, 0, 5, 10]">
+        <atft-object-loader *ngFor="let x of [0,5,10,15,20]" model="assets/model/Server.json" [translateX]="x" [translateZ]="z"
+          [rotateX]="rotationX"
+          [rotateY]="rotationY"
+          [rotateZ]="rotationZ"
+          [translateY]="translationY"
+          >
+        </atft-object-loader>
+      </div>
+
+      <atft-empty translateZ=-50>
+        <atft-sphere-mesh atft-raycaster-group (mouseEnter)="mouseEnter()" (mouseExit)="mouseExit()" (click)="click()"
+                          [radius]="4" [widthSegments]="20" [hightSegments]="20" material="phong" materialColor="0xff0000"
+                          [translateX]=10 [translateY]=5 [translateZ]="20">
+        </atft-sphere-mesh>
+        <atft-cylinder-mesh [radiusTop]="2" [radiusBottom]="3" [height]="10" [radialSegments]="36" [heightSegments]="1"
+          material="lamb" materialColor="0x00ff00" [translateY]="5" [translateZ]="10">
+        </atft-cylinder-mesh>
+        <atft-torus-mesh [radius]="5" [tube]="2" [radialSegments]="16" [tubularSegments]="100" [arc]="6.28" material="lamb"
+          materialColor="0x0000ff" [rotateX]="1.6" [translateX]="-10" [translateY]="2" [translateZ]="-5">
+        </atft-torus-mesh>
+      </atft-empty>
+
+    </atft-scene>
+  </atft-renderer-canvas>
+</atft-orbit-controls>
