+++ src/app/app.component.scss
@@ -1,0 +1,6 @@
+:host {
+  width: 100%;
+  height: 100%;
+  max-height: 100%;
+  display: flex;
+}
