+++ src/app/app.component.ts
@@ -6,5 +6,25 @@ import { Component } from '@angular/core';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
+  title = 'app';
 
+  public rotationX = 0.0;
+  public rotationY = 0.0;
+  public rotationZ = 0.0;
+
+  public translationY = 0.0;
+
+  public cameraPositionX = 20.0;
+
+  public mouseEnter() {
+    console.log('mouseEnter');
+  }
+
+  public click() {
+    console.log('click');
+  }
+
+  public mouseExit() {
+    console.log('mouseExit');
+  }
 }
