+++ src/app/app.module.ts
@@ -1,10 +1,11 @@
-import {NgModule} from '@angular/core';
-import {FormsModule} from '@angular/forms';
 import {BrowserModule} from '@angular/platform-browser';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {AppComponent} from './app.component';
-import {EditorModule} from './editor/editor.module';
+import {FormsModule} from '@angular/forms';
+import {NgModule} from '@angular/core';
 
+import {AppComponent} from './app.component';
+import {AtftModule} from 'atft';
+
+// For development (code watch): import { AtftModule } from 'projects/atft/src/lib/atft.module';
 
 @NgModule({
   declarations: [
@@ -13,11 +14,9 @@ import {EditorModule} from './editor/editor.module';
   imports: [
     BrowserModule,
     FormsModule,
-    BrowserAnimationsModule,
-    EditorModule
+    AtftModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule {
-}
+export class AppModule { }
