--- src/app/editor/editor-canvas/editor-canvas.component.html
@@ -1,49 +1,0 @@
-<atft-renderer-canvas>
-  <atft-orthographic-camera [positionX]=15 [positionY]=-50 [positionZ]="100" [zoom]="6" [zAxisUp]="true">
-  </atft-orthographic-camera>
-
-  <atft-scene name="scene" background="0xa0a0a0">
-
-    <atft-plane-mesh name="ground" height="2000" width="2000" materialColor="0x999999" [depthWrite]="true" [castShadow]="false"
-                     material="phong" [receiveShadow]="false" heightSegments="1" widthSegments="1" translateZ="-0.1">
-    </atft-plane-mesh>
-
-    <atft-hemisphere-light name="hemi-light" skyColor="0xffffff" groundColor="0xffffff" intensity="0.4"
-                           translateX="-20" translateY="-20" translateZ="100">
-    </atft-hemisphere-light>
-
-    <atft-point-light name="point-light" intensity="0.8"
-                      translateX="-100" translateY="-100" translateZ="200" [castShadow]="true">
-    </atft-point-light>
-
-
-    <atft-dagre-layout>
-
-      <atft-dagre-node name="spa">
-        <atft-server-compact-actor label="SPA"></atft-server-compact-actor>
-      </atft-dagre-node>
-
-      <atft-dagre-node name="api">
-        <atft-server-stand-actor label="API"></atft-server-stand-actor>
-      </atft-dagre-node>
-      <atft-dagre-node name="kv">
-        <atft-server-compact-actor label="Key Vault"></atft-server-compact-actor>
-      </atft-dagre-node>
-      <atft-dagre-node name="db1" composition="data">
-        <atft-server-barrel-actor label="PostgreSQL"></atft-server-barrel-actor>
-      </atft-dagre-node>
-      <atft-dagre-node name="db2" composition="data">
-        <atft-server-barrel-actor label="MongoDB"></atft-server-barrel-actor>
-      </atft-dagre-node>
-
-      <atft-dagre-edge from="spa" to="api"></atft-dagre-edge>
-      <atft-dagre-edge from="api" to="db1"></atft-dagre-edge>
-      <atft-dagre-edge from="api" to="db2"></atft-dagre-edge>
-      <atft-dagre-edge from="api" to="kv"></atft-dagre-edge>
-
-      <atft-dagre-composition name="data" label="Data"></atft-dagre-composition>
-
-    </atft-dagre-layout>
-
-  </atft-scene>
-</atft-renderer-canvas>
