--- src/app/editor/editor-canvas/editor-canvas.component.scss
@@ -1,3 +1,0 @@
-:host {
-  flex: 1 0 auto;
-}
