--- src/app/editor/editor-content/editor-content.component.html
@@ -1,8 +1,0 @@
-<mat-drawer-container class="example-container" [hasBackdrop]="false">
-  <mat-drawer mode="push" position="end">
-    <app-editor-side></app-editor-side>
-  </mat-drawer>
-  <mat-drawer-content>
-    <app-editor-canvas></app-editor-canvas>
-  </mat-drawer-content>
-</mat-drawer-container>
