--- src/app/editor/editor-content/editor-content.component.scss
@@ -1,19 +1,0 @@
-:host {
-  flex: 1 0 auto;
-  display: flex;
-}
-
-.mat-drawer-container {
-  flex: 1 1 auto;
-  display: flex;
-}
-
-.mat-drawer-content {
-  flex: 1 0 auto;
-  display: flex;
-}
-
-.mat-drawer-push {
-  width: 10%;
-  min-width: 150px;
-}
