--- src/app/editor/editor-content/editor-content.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { AnimationService } from 'atft';
-
-@Component({
-  selector: 'app-editor-content',
-  templateUrl: './editor-content.component.html',
-  styleUrls: ['./editor-content.component.scss']
-})
-export class EditorContentComponent {
-
-}
