--- src/app/editor/editor-layout/editor-layout.component.html
@@ -1,2 +1,0 @@
-<app-editor-toolbar></app-editor-toolbar>
-<app-editor-content></app-editor-content>
