--- src/app/editor/editor-layout/editor-layout.component.scss
@@ -1,7 +1,0 @@
-:host {
-  width: 100%;
-  height: 100%;
-  max-height: 100%;
-  display: flex;
-  flex-flow: column;
-}
