--- src/app/editor/editor-layout/editor-layout.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-editor-layout',
-  templateUrl: './editor-layout.component.html',
-  styleUrls: ['./editor-layout.component.scss']
-})
-export class EditorLayoutComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
