--- src/app/editor/editor-side/editor-side.component.scss
@@ -1,2 +1,0 @@
-
-
