--- src/app/editor/editor-side/editor-side.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-editor-side',
-  templateUrl: './editor-side.component.html',
-  styleUrls: ['./editor-side.component.scss']
-})
-export class EditorSideComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
