--- src/app/editor/editor-toolbar/editor-toolbar.component.html
@@ -1,7 +1,0 @@
-<mat-toolbar>
-  <span>Editor</span>
-  <span class="example-spacer"></span>
-  <button mat-icon-button aria-label="Edit">
-    <mat-icon>create</mat-icon>
-  </button>
-</mat-toolbar>
