--- src/app/editor/editor-toolbar/editor-toolbar.component.scss
@@ -1,8 +1,0 @@
-:host {
-  flex: 0 1 auto;
-}
-
-.example-spacer {
-  flex: 1 1 auto;
-}
-
