--- src/app/editor/editor-toolbar/editor-toolbar.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { EditorToolbarComponent } from './editor-toolbar.component';
-
-describe('EditorToolbarComponent', () => {
-  let component: EditorToolbarComponent;
-  let fixture: ComponentFixture<EditorToolbarComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ EditorToolbarComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EditorToolbarComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
