--- src/app/editor/editor-toolbar/editor-toolbar.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-editor-toolbar',
-  templateUrl: './editor-toolbar.component.html',
-  styleUrls: ['./editor-toolbar.component.scss']
-})
-export class EditorToolbarComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
